pyproject.toml
[project]
name = "openaitool"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "geopy>=2.4.1",
    "openai>=1.59.8",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
]
